+++ src/main.js
@@ -14,6 +14,7 @@ import { extendStoreWithScheme } from '@/vuex'
 import { buildRouter } from '@/vue-router'
 import { tableScrollShadow } from '@/vue/directives/tableScrollShadow'
 import { ripple } from '@/vue/directives/ripple'
+import { tooltip } from '@/vue/directives/tooltip'
 import { i18nOptions, mergeIntoI18nOptions } from '@/i18n'
 import { globalize } from '@/vue/filters/globalize'
 import { formatDate } from '@/vue/filters/formatDate'
@@ -47,6 +48,7 @@ async function init () {
   Vue.use(VueResource)
   Vue.directive('table-scroll-shadow', tableScrollShadow)
   Vue.directive('ripple', ripple)
+  Vue.directive('tooltip', tooltip)
   Vue.filter('globalize', globalize)
   Vue.filter('formatDate', formatDate)
   Vue.filter('formatDateDMY', formatDateDMY)