+++ src/vuex/account.module.js
@@ -26,20 +26,25 @@ export const actions = {
     commit(vuexTypes.SET_ACCOUNT, response.data)
   },
 
-  async [vuexTypes.LOAD_ACCOUNT_BALANCES_DETAILS] ({ commit, getters }) {
+  async [vuexTypes.LOAD_ACCOUNT_BALANCES_DETAILS] (
+    { commit, rootGetters, getters }
+  ) {
     const accountId = getters[vuexTypes.accountId]
-    const endpoint = `/v3/accounts/${accountId}`
-    const { data: account } = await api.getWithSignature(endpoint, {
-      include: ['balances.asset', 'balances.state'],
+    const defaultQuoteAsset = rootGetters[vuexTypes.defaultQuoteAsset]
+
+    const endpoint = `/v3/accounts/${accountId}/converted_balances/${defaultQuoteAsset}`
+    const { data } = await api.getWithSignature(endpoint, {
+      include: ['states', 'balance', 'balance.state', 'balance.asset'],
     })
+
+    const balances = data.states.map(state => state.balance)
 
     commit(
       vuexTypes.UPDATE_ASSETS,
-      account.balances.map(b => b.asset),
+      balances.map(b => b.asset),
       { root: true }
     )
-
-    commit(vuexTypes.SET_ACCOUNT_BALANCES_DETAILS, account.balances)
+    commit(vuexTypes.SET_ACCOUNT_BALANCES_DETAILS, balances)
   },
 }
 
@@ -47,8 +52,7 @@ export const getters = {
   [vuexTypes.account]: state => state.account,
   [vuexTypes.accountId]: state => state.account.id,
   [vuexTypes.accountBalances]: state => state.balancesDetails
-    .map(item => new BalanceRecord(item))
-    .sort((a, b) => b.convertedBalance - a.convertedBalance),
+    .map(item => new BalanceRecord(item)),
   [vuexTypes.accountRoleId]: state => Number(
     _get(state.account, 'role.id')
   ),