+++ src/vuex/account.spec.js
@@ -51,6 +51,7 @@ describe('account.module', () => {
         getters: {
           accountId: 'GAIEBMXUPSGW2J5ELJFOY6PR5IWXXJNHIJSDKTDHK76HHRNYRL2QYU4O',
         },
+        rootGetters: {},
         commit: sinon.stub(),
         dispatch: sinon.stub(),
       }
@@ -72,10 +73,12 @@ describe('account.module', () => {
 
     it('LOAD_ACCOUNT_BALANCES_DETAILS commits proper set of mutations',
       async () => {
-        sinon.stub(api, 'getWithSignature').resolves({
-          data: { balances: balancesDetailsJSON },
-        })
-        const assetsPayload = balancesDetailsJSON.map(b => b.asset)
+        const balancesMock = MockWrapper
+          .makeJsonapiResponseData(balancesDetailsJSON)
+        sinon.stub(api, 'getWithSignature').resolves({ data: balancesMock })
+
+        const balances = balancesMock.states.map(state => state.balance)
+        const assetsPayload = balances.map(b => b.asset)
 
         await actions[vuexTypes.LOAD_ACCOUNT_BALANCES_DETAILS](store)
 
@@ -83,7 +86,7 @@ describe('account.module', () => {
           vuexTypes.UPDATE_ASSETS, assetsPayload, { root: true }
         )
         expect(store.commit).to.have.been.calledWithExactly(
-          vuexTypes.SET_ACCOUNT_BALANCES_DETAILS, balancesDetailsJSON
+          vuexTypes.SET_ACCOUNT_BALANCES_DETAILS, balances
         )
       })
   })